+++ app/src/main/java/com/dntks/groupstagesimulator/data/model/RoundDomainModel.kt
@@ -1,7 +1,7 @@
 package com.dntks.groupstagesimulator.data.model
 
 data class RoundDomainModel(
-    val roundId: Int,
+    val roundId: Long,
     val roundName: String,
     val matches: List<MatchDomainModel>
 )
