+++ app/src/main/java/com/dntks/groupstagesimulator/data/model/TeamDomainModel.kt
@@ -1,7 +1,7 @@
 package com.dntks.groupstagesimulator.data.model
 
 data class TeamDomainModel(
-    val teamId: Int,
+    val teamId: Long,
     val name: String,
     val players: List<PlayerDomainModel>
